+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/manageaccount/backupconfirmkey/BackupConfirmKeyFragment.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.backupconfirmkey
+package io.horizontalsystems.bankwallet.modules.manageaccount.backupconfirmkey
 
 import android.os.Bundle
 import android.view.LayoutInflater
